add repair complete function

+++ app/Scripts/php/RssScript.php
@@ -42,7 +42,7 @@ class RssScript{
     }
 
     public function RssToDB($count = 0){
-        $this->log('####BEGIN SCRIPT####');
+        $this->log('###BEGIN SCRIPT####');
         try {
             $feeds = Data::feeds();
             $totalDuration = -microtime(true);
@@ -52,12 +52,12 @@ class RssScript{
                 $durationFeed = -microtime(true);
                 $newArticles = 0;
                 $newRelations = 0;
-                $this->log("####Feed : $url");//LOG
+                $this->log("###Feed : $url");//LOG
                 // Va chercher les feeds
                 try {
                     $feeds = FeedIo::create()->getFeedIo()->read($url)->getFeed();
                 } catch (\Throwable $e) {
-                    $this->log("###CURL exception on $url : {$e->getMessage()}");
+                    $this->log("####CURL exception on $url : {$e->getMessage()}");
                     $feeds = [];
                 }
                 // Parcours les articles d'un flux rss
@@ -83,12 +83,12 @@ class RssScript{
                         }
                     }
                 } catch (\Throwable $e) {
-                    $this->log("###XML exception on $url : {$e->getMessage()}");
+                    $this->log("####XML exception on $url : {$e->getMessage()}");
                 }
                 $durationFeed += microtime(true);
-                $this->log("####Articles : $newArticles");//LOG
-                $this->log("####Relations : $newRelations");//LOG
-                $this->log("####Duration : $durationFeed");//LOG
+                $this->log("###Articles : $newArticles");//LOG
+                $this->log("###Relations : $newRelations");//LOG
+                $this->log("###Duration : $durationFeed");//LOG
                 $relationTotal+=$newRelations;
                 $newArticlesTotal+=$newArticles;
             }
@@ -99,23 +99,29 @@ class RssScript{
             $this->log("##Total Duration : $totalDuration");//LOG
             
         } catch (\Throwable $e) {
-            $this->log($e);
+            $this->log("####$e");
             $this->log($count);
             if ($count < 5) {
-                // $this->RssToDB($count ++);
+                $this->RssToDB($count ++);
             }
         }
     }
 
-    public function getArticle($feed){
-        $article = new Article;
-        $mediaName = explode('.',isset(parse_url( $feed->getLink())['host'])?parse_url($feed->getLink())['host']:$feed->getLink())[1];
-        $article->media = ($mediaName == "google")?"rtl":$mediaName;
-        $article->titre = $feed->getTitle();
-        $article->description = strip_tags($feed->getDescription());
-        $article->date = $feed->getLastModified()->format('Y-m-d H:i:s');
-        $article->lien = $feed->getLink();
-        $article->article = ($article->medi=="rtl")?null:$this->getCompleteArticle($article->lien, $article->media); // pour rtl la description contient tout l'article
+    public function getArticle($feed, $update = ["check"=>false, "id"=>""]){
+        $article = ($update['check'])? Article::find($update['id']) : new Article;
+
+        if (!$update['check']) {
+            $mediaName = explode('.',isset(parse_url( $feed->getLink())['host'])?parse_url($feed->getLink())['host']:$feed->getLink())[1];
+            $article->media = ($mediaName == "google")?"rtl":$mediaName;
+            $article->titre = $feed->getTitle();
+            $article->description = strip_tags($feed->getDescription());
+            $article->date = $feed->getLastModified()->format('Y-m-d H:i:s');
+            $article->lien = $feed->getLink();
+            $article->article = ($article->media=="rtl")?null:$this->getCompleteArticle($article->lien, $article->media); // pour rtl la description contient tout l'article
+        }else{
+            $article->article = ($article->media=="rtl")?null:$this->getCompleteArticle($article->lien, $article->media);
+            $article->number_testing++;
+        }
         $article->save();
         return $article;
     }
@@ -125,14 +131,17 @@ class RssScript{
             'removeDoubleSpace'=>true
         ]);
         $dom->loadFromUrl($url);
-        $tag = Data::mediasTag()[$media];
-        $content = $dom->find($tag);
         $data = "";
-        foreach($content as $e){
-            foreach ($e->find('a') as $a){
-                $a->delete();
+        foreach (Data::mediasTag()[$media] as $tag) {
+            if($data == ""){
+                $content = $dom->find($tag);
+                foreach($content as $e){
+                    foreach ($e->find('a') as $a){
+                        $a->delete();
+                    }
+                    $data .= $e->innerHtml;
+                }
             }
-            $data .= $e->innerHtml;
         }
         return html_entity_decode(trim(strip_tags(htmlspecialchars_decode($data, ENT_QUOTES))));
     }
@@ -148,29 +157,48 @@ class RssScript{
 
     public function getPoliticianInformation($politician, $id){
         $description="";$descriptionLink="";$imageLink="";
-        try {
-            // get information
-            $content = API::wikipedia($politician, "content");
-            $description = (isset($content[2][0]) && $content[2][0] != '')? $content[2][0] : null;
-            $descriptionLink = (isset($content[3][0]) && $content[3][0] != '' ) ? $content[3][0] : null;
-            
-            // Get image url
-            $image = API::wikipedia($politician, "image");
-            $imageLink = (!empty($image['query']['pages'][0]))? ((!empty($image['query']['pages'][0]['thumbnail']['source']))? $image['query']['pages'][0]['thumbnail']['source']: null) : null;        
-        } catch (\Throwable $th) {
-            $this->log("###ERROR API - WIKIPEDIA :{$th->getMessage()}");
+        if(Politician::where('id',$id)->where('number_testing','<', 5)->first()){
+            try {
+                // get information
+                $content = API::wikipedia($politician, "content");
+                $description = (isset($content[2][0]) && $content[2][0] != '')? $content[2][0] : null;
+                $descriptionLink = (isset($content[3][0]) && $content[3][0] != '' ) ? $content[3][0] : null;
+                
+                // Get image url
+                $image = API::wikipedia($politician, "image");
+                $imageLink = (!empty($image['query']['pages'][0]))? ((!empty($image['query']['pages'][0]['thumbnail']['source']))? $image['query']['pages'][0]['thumbnail']['source']: null) : null;        
+            } catch (\Throwable $th) {
+                $this->log("####ERROR API - WIKIPEDIA :{$th->getMessage()}");
+            }
+            // INSERT
+            try {
+                $politician= Politician::find($id);
+                $politician->description = $description;
+                $politician->lienDescription = $descriptionLink;
+                $politician->image = $imageLink;
+                $politician->number_testing++;
+                $politician->save();
+            } catch (\Throwable $th) {
+                $this->log("####ERROR SQL INSERT POLITICIAN INFORMATIONS :{$th->getMessage()}");
+            }
         }
 
-        // INSERT
-        try {
-            $politician= Politician::find($id);
-            $politician->description = $description;
-            $politician->lienDescription = $descriptionLink;
-            $politician->image = $imageLink;
-            $politician->save();
-        } catch (\Throwable $th) {
-            $this->log("###ERROR SQL INSERT POLITICIAN INFORMATIONS :{$th->getMessage()}");
+    }
+
+    public function getPoliticianCitationsByArticle($lastname, $firstname){
+        $regex = '/\..*[marghem]?.*("([^"]|"")*").*[marghem]?.*\./';
+
+    }
+
+    public function repairCompleteAll(){
+        // ARTICLES GET COMPLETE
+        $this->log("##INCOMPLETE ARTICLES");
+        $incompleteArticles = Article::where('article', "=", "")->where('number_testing', '<', '5')->get();
+        foreach ($incompleteArticles as $article) {
+            $this->log("##TRYING UPDATE ARTICLE NUMBER $article->id");
+            $this->getArticle("", ['check'=>true, "id"=>$article->id]);
         }
+
     }
 
 }
